chore: consume from weather API

+++ weatherapi/serializers.py
@@ -3,8 +3,8 @@ from rest_framework import serializers
 
 class WeatherSerializer(serializers.Serializer):
     """The weather result serializer, define your fields here."""
-    maximum = serializers.IntegerField()
-    minimum = serializers.IntegerField()
-    average = serializers.IntegerField()
-    median = serializers.IntegerField()
+    maximum = serializers.FloatField()
+    minimum = serializers.FloatField()
+    average = serializers.FloatField()
+    median = serializers.FloatField()
